# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Le Kiosque Pédagogique API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets/create:
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a new anonymous ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Ticket created with credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketCreated"
  /tickets/login:
    post:
      operationId: loginTicket
      tags: [tickets]
      summary: Authenticate as an anonymous teacher using ticket number + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketCredentials"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid ticket credentials
  /tickets/me:
    get:
      operationId: getMyTicket
      tags: [tickets]
      summary: Get the authenticated teacher's ticket
      responses:
        "200":
          description: Ticket detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
        "401":
          description: Unauthorized
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get a ticket by ID (staff only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
        "403":
          description: Forbidden
        "404":
          description: Not found
  /tickets/{id}/claim:
    patch:
      operationId: claimTicket
      tags: [tickets]
      summary: Claim a ticket (N1 or N2)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
        "409":
          description: Already claimed
  /tickets/{id}/escalate:
    patch:
      operationId: escalateTicket
      tags: [tickets]
      summary: Escalate a ticket to N2 (N1 only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket escalated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
  /tickets/{id}/resolve-n1:
    patch:
      operationId: resolveTicketN1
      tags: [tickets]
      summary: Resolve a ticket at level 1 (N1 only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
  /tickets/{id}/resolve:
    patch:
      operationId: resolveTicket
      tags: [tickets]
      summary: Resolve a ticket at level 2 (N2 only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
  /tickets/{id}/webex:
    patch:
      operationId: closeTicketWebex
      tags: [tickets]
      summary: Schedule a collective visio session (F3 only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebexInput"
      responses:
        "200":
          description: Ticket closed with Webex
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
  /tickets/{id}/reassign-n1:
    patch:
      operationId: reassignTicketN1
      tags: [tickets]
      summary: Reassign a ticket to another F2 (current assignee only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReassignInput"
      responses:
        "200":
          description: Ticket reassigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
        "409":
          description: Cannot reassign in current state
  /tickets/{id}/reassign-n2:
    patch:
      operationId: reassignTicketN2
      tags: [tickets]
      summary: Reassign a ticket to another F3 (current assignee only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReassignInput"
      responses:
        "200":
          description: Ticket reassigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
        "409":
          description: Cannot reassign in current state

  # ── MESSAGES ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AuthResult:
      type: object
      required: [token]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, role]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        schoolId:
          type: ["integer", "null"]
        disciplineId:
          type: ["integer", "null"]
        active:
          type: boolean
        school:
          $ref: "#/components/schemas/School"
        discipline:
          $ref: "#/components/schemas/Discipline"
...
    TicketInput:
      type: object
      required: [schoolId, disciplineId, description]
      properties:
        schoolId:
          type: integer
        disciplineId:
          type: integer
        description:
          type: string
    TicketCredentials:
      type: object
      required: [ticketNumber, password]
      properties:
        ticketNumber:
          type: integer
        password:
          type: string
    TicketCreated:
      type: object
      required: [ticketNumber, password, token]
      properties:
        ticketNumber:
          type: integer
        password:
          type: string
        token:
          type: string
...
    TicketDetail:
      type: object
      required: [id, status, createdAt]
      properties:
        id:
          type: integer
        status:
          type: string
        schoolId:
          type: integer
        disciplineId:
          type: integer
        description:
          type: ["string", "null"]
        webexLink:
          type: ["string", "null"]
        webexScheduledAt:
          type: ["string", "null"]
          format: date-time
        webexCreatedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        school:
          $ref: "#/components/schemas/School"
        discipline:
          $ref: "#/components/schemas/Discipline"
        assignedN1Id:
          type: ["integer", "null"]
          description: Visible par le staff uniquement
        assignedN2Id:
          type: ["integer", "null"]
          description: Visible par le staff uniquement
...
    ReassignInput:
      type: object
      required: [targetUserId]
      properties:
        targetUserId:
          type: integer
    WebexInput:
      type: object
      required: [webexLink, scheduledAt]
      properties:
        webexLink:
          type: string
        scheduledAt:
          type: string
          format: date-time
          description: Date et heure de la session visio

    # ── MESSAGES ─────────────────────────────────────────────────────────
...
    School:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: ["string", "null"]
        active:
          type: boolean
...
    Discipline:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        active:
          type: boolean
...