# lib/api-zod/src/generated/types/ticketDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Le Kiosque Pédagogique API
 * OpenAPI spec version: 0.1.0
 */
import type { Discipline } from './discipline';
import type { School } from './school';

export interface TicketDetail {
  id: number;
  status: string;
  schoolId?: number;
  disciplineId?: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  webexLink?: string | null;
  /** @nullable */
  webexScheduledAt?: Date | null;
  /** @nullable */
  webexCreatedAt?: string | null;
  createdAt: string;
  updatedAt?: string;
  school?: School;
  discipline?: Discipline;
  /**
     * Visible par le staff uniquement
     * @nullable
     */
  assignedN1Id?: number | null;
  /**
     * Visible par le staff uniquement
     * @nullable
     */
  assignedN2Id?: number | null;
}